backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
django-cors-headers>=4.4,<4.8
psycopg[binary]>=3.2,<3.3
